new home page now styled and has link to bagels page
